requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.4
Django==1.8.5
django-bootstrap3-datepicker==0.4.1
django-bootstrap3-datetimepicker-2==2.4.2
django-braces==1.11.0
django-crispy-forms==1.6.1
django-input-mask==2.0.2
django-oauth-toolkit==1.0.0
django-suit==0.2.25
idna==2.5
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
pytz==2017.2
reportlab==3.4.0
requests==2.18.1
six==1.10.0
Unipath==1.1
urllib3==1.21.1
